refactor admin routes to a folder
