+++ backend/app/routes/admin/module.route.js
@@ -1,13 +1,13 @@
 const express = require("express");
 const router = express.Router();
-const ModuleService = require("../services/module.services");
-const ModuleController = require("../controllers/module.controller");
-const ValidateTokenMiddleware = require("../middleware/validateToken.middleware");
-const BlackListService = require("../services/blacklist.service");
-const ValidateRoleMiddleware = require("../middleware/validateRole.middleware");
-const Role = require("../enums/roles.enum");
-const {validate} = require("../middleware/validate.middleware");
-const {validateListModules, validateModuleIndex} = require("../validators/module.validator");
+const ModuleService = require("../../services/module.services");
+const ModuleController = require("../../controllers/module.controller");
+const ValidateTokenMiddleware = require("../../middleware/validateToken.middleware");
+const BlackListService = require("../../services/blacklist.service");
+const ValidateRoleMiddleware = require("../../middleware/validateRole.middleware");
+const Role = require("../../enums/roles.enum");
+const {validate} = require("../../middleware/validate.middleware");
+const {validateListModules, validateModuleIndex} = require("../../validators/module.validator");
 
 const validateTokenMiddleware = new ValidateTokenMiddleware(new BlackListService());
 const validateRoleMiddleware = new ValidateRoleMiddleware();
