+++ backend/app/routes/admin/user.route.js
@@ -1,20 +1,20 @@
 const express = require('express');
 const router = express.Router();
 
-const {validateUserRegistration} =require('../validators/user.validator');
-const {validate} = require("../middleware/validate.middleware");
-const UserController = require('../controllers/user.controller');
-const ValidateTokenMiddleware = require('../middleware/validateToken.middleware');
-const BlackListService = require('../services/blacklist.service');
-const UserService = require("../services/user.service");
-const Mailer = require('../utils/Mailer');
+const {validateUserRegistration} =require('../../validators/user.validator');
+const {validate} = require("../../middleware/validate.middleware");
+const UserController = require('../../controllers/user.controller');
+const ValidateTokenMiddleware = require('../../middleware/validateToken.middleware');
+const BlackListService = require('../../services/blacklist.service');
+const UserService = require("../../services/user.service");
+const Mailer = require('../../utils/Mailer');
 const validateTokenMiddleware = new ValidateTokenMiddleware(new BlackListService());
 const mailer = new Mailer(process.env.RESEND_API_KEY);
 const userService = new UserService(mailer);
 const userController = new UserController(userService);
-const Role = require("../enums/roles.enum");
-const ValidateRoleMiddleware = require("../middleware/validateRole.middleware");
-const ValidateUserCreationMiddleware = require("../middleware/validateUserCreation.middleware");
+const Role = require("../../enums/roles.enum");
+const ValidateRoleMiddleware = require("../../middleware/validateRole.middleware");
+const ValidateUserCreationMiddleware = require("../../middleware/validateUserCreation.middleware");
 
 const validateRoleMiddleware = new ValidateRoleMiddleware();
 const validateUserCreation = new ValidateUserCreationMiddleware();
